requirements.txt
cvxopt==1.3.0
glcontext==2.3.6
moderngl==5.8.2
multipledispatch==0.6.0
numpy==1.23.4
Pillow==9.4.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt5-stubs==5.15.6.0
pyrr==0.10.3
six==1.16.0
style==1.1.0
update==0.0.1
util-moderngl-qt==0.1.19
del-msh==0.1.11
del-srch==0.1.4
